Added: uploads real data to firebase
TODO:
save as CSV still doesnt work

+++ app/src/main/java/edu/uco/cmanley1/accelerometersensor/CoordinateSet.kt
@@ -2,6 +2,6 @@ package edu.uco.cmanley1.accelerometersensor
 
 import java.sql.Timestamp
 
-class CoordinateSet(var x: Double, var y: Double, var z: Double, var timestamp: Timestamp)
+class CoordinateSet(var x: Float, var y: Float, var z: Float, var timestamp: Timestamp)
 {
 }
